Migrations, seeders y readme
I tried cloning the project from scratch on another computer to check that the instructions in the readme file work correctly and that the application can be tested.
I added the migrations (I had created them in a previous test attempt!) and seeders so that you can interact with the application a little.